==> notebooks/06_statistiques_descriptives.ipynb <==
# Etape 3.1 : Statistiques descriptives

import pandas as pd

df = pd.read_parquet("../output/consommations_enrichies.parquet")

df.shape, df.head()


## Calculer les statistiques par type d'énergie, type de bâtiment et commune

# apr type d'énergie 
stats_energie = (
    df
    .groupby("type_energie")["conso_clean"]
    .agg(["count", "sum", "mean", "median", "std"])
    .reset_index()
)

stats_energie
# export 
stats_energie.to_csv("../output/stats_par_type_energie.csv", index=False)

# apr type de bâtiment 

stats_type_bat = (
    df
    .groupby("type")["conso_par_m2"]
    .agg(["count", "mean", "median", "std"])
    .reset_index()
    .sort_values("mean", ascending=False)
)

stats_type_bat

stats_type_bat.to_csv("../output/stats_par_type_batiment.csv", index=False)

# par cmmune 

stats_commune = (
    df
    .groupby("commune")["conso_clean"]
    .sum()
    .reset_index(name="conso_totale_kwh")
    .sort_values("conso_totale_kwh", ascending=False)
)

stats_commune

stats_commune.to_csv("../output/stats_par_commune.csv", index=False)


## Identifier les batiments les plus/moins energivores

conso_bat = (
    df
    .groupby(["batiment_id", "nom", "type", "commune"])["conso_par_m2"]
    .mean()
    .reset_index()
)

# top 10 les plus enegivores
top_plus = conso_bat.sort_values("conso_par_m2", ascending=False).head(10)
top_plus

# top 10 les moins enegivores
top_moins = conso_bat.sort_values("conso_par_m2", ascending=True).head(10)
top_moins

# export 
top_plus.to_csv("../output/batiments_plus_energivores.csv", index=False)
top_moins.to_csv("../output/batiments_moins_energivores.csv", index=False)


## Calculer la repartition des consommations par classe energetique DPE

repartition_dpe = (
    df
    .groupby("classe_energetique")["conso_clean"]
    .sum()
    .reset_index(name="conso_totale_kwh")
    .sort_values("classe_energetique")
)

repartition_dpe
# option %

repartition_dpe["part_%"] = (
    repartition_dpe["conso_totale_kwh"]
    / repartition_dpe["conso_totale_kwh"].sum()
    * 100
)

repartition_dpe.to_csv("../output/repartition_conso_dpe.csv", index=False)


## Analyser l'evolution temporelle (tendances mensuelles, saisonnalite)

# mensuel 
df["mois"] = pd.to_datetime(df["date_jour"]).dt.to_period("M")

conso_mensuelle = (
    df
    .groupby("mois")["conso_clean"]
    .sum()
    .reset_index()
)

conso_mensuelle
conso_mensuelle.to_csv("../output/conso_mensuelle.csv", index=False)

# saison 

conso_saison = (
    df
    .groupby("saison")["conso_clean"]
    .sum()
    .reset_index()
)

conso_saison
conso_saison.to_csv("../output/conso_par_saison.csv", index=False)


## Comparer la consommation theorique (selon DPE) vs reelle

dpe_ref = {
    "A": 50,
    "B": 90,
    "C": 150,
    "D": 230,
    "E": 330,
    "F": 420,
    "G": 500
}

df["conso_theorique_kwh_m2"] = df["classe_energetique"].map(dpe_ref)


# comparaison 
comparaison_dpe = (
    df
    .groupby("classe_energetique")
    .agg(
        conso_reelle_moy=("conso_par_m2", "mean"),
        conso_theorique=("conso_theorique_kwh_m2", "mean")
    )
    .reset_index()
)

comparaison_dpe["ecart_reel_vs_theorique"] = (
    comparaison_dpe["conso_reelle_moy"]
    - comparaison_dpe["conso_theorique"]
)

comparaison_dpe

# export
comparaison_dpe.to_csv("../output/comparaison_dpe_reel_vs_theorique.csv", index=False)
